// src/app/models/employee-atk.ts
export class EmployeeATK {
    ContactId?: number;
    EmployeeId?: number;
    FirstName?: number;
    LastName?: number;
    Company?: number;
}

export class Employee {
    contactId?: number;
    employeeId?: number;
    firstName?: number;
    lastName?: number;
    company?: number;
}

export class UploadFile {
    fileId?: number;
    fileName?: string;
}